Improved color tracking and rectangles drawing.
body: rectangles with a perimeter longer than a certain amount are drawn around the color detected over the same original frame recorded and then shown. Now the smaller and useless rectangles are not considered at all.
issues: currently works fine for stronger greens detected on the laptop cam, to be adapted for lighter and weaker greens of the stickers on the car, and tested on the replays