Improved Hue green detection
body:
More correct detection of the stickers colo
Perimeter and area of the detected rectangles are printed in console for feedback on the smallest interesting values of detection.

issues:
It detects ALL the rectangles with a perimeter longer than an arbitrary value.
Needs to be improved with limiting the rectangles to 3, or to merge all the rectangles too close to each other.
Cannot blur the image more than size(1,1), probably due to the small video resolution.
To be tested on different color scenarios due to new HSV values.